requirements.txt
Cython==3.0.11
Jinja2==3.1.6
joblib==1.4.2
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.16.0
mkl-service==2.4.0
opt_einsum==3.4.0
osqp==1.0.1
platformdirs==4.3.6
py-spy==0.4.0
Pygments==2.19.1
rich==13.9.4
scikit-learn==1.6.0
scs==3.2.7.post2
sdpa-python==0.2.2
textual==2.1.2
threadpoolctl==3.5.0
typing_extensions==4.12.2
uc-micro-py==1.0.3
cvxpy==1.6.4
numpy==2.1.3
PyYAML==6.0.2
scipy==1.14.1
